--- a/torch_tracker.py
+++ b/torch_tracker.py
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: 2022-present deepset GmbH <info@deepset.ai>
+#
+# SPDX-License-Identifier: Apache-2.0
+
 import importlib.abc
 import importlib.util
 import sys