### Motivation

Motivated by recent issues (#8649, #8650, and #8704), I investigated the import times of Haystack.

I used some raw scripts (attached in this draft PR) and commands like
`python -X importtime -c "import haystack"` to analyze cumulative import times of packages.

### Findings
- Our `LazyImport` context manager behaves like a `DeferImportError` context manager. If a package you import is available, it gets imported immediately. If not, the import error is caught and only raised later when calling  `.check()`. Maybe this is something known, but it was surprising to me.
- As a result the more libraries installed, the slower the imports (since more packages get loaded).
- Our telemetry surprisingly imports heavy dependencies like `torch` and `transformers`.

### An exploratory solution

I explored low-effort ways to improve this:
- remove heavy imports from telemetry
- make modules import in `__init__.py` of packages *actually lazy*
  - I only did this for the `utils` package, that triggered the `torch` import

#### Results

Tested on Ubuntu 22.04, Python 3.10.12, averaged over 30 runs.

I focus on User time and System time, as they capture actual CPU time, irrespective of parallelism and system load ([reference](https://stackoverflow.com/questions/556405/what-do-real-user-and-sys-mean-in-the-output-of-time1)).

*Dependencies*
- Minimal = `pip install haystack-ai`
- Test = All dependencies required for tests ([pyproject](https://github.com/deepset-ai/haystack/blob/e5b9bdeb66b8e459a269a91ad922c6045dc2ad79/pyproject.toml#L85))

| Haystack | Dependencies   | User Time | Sys Time | Total Time |
|----------|--------|-----------|----------|------------|
| Current (https://github.com/deepset-ai/haystack/commit/35788a2d060d58b901b8d10237f23273399998e5)  | Minimal | 2.114s    | 0.031s   | 2.145s     |
| Current (https://github.com/deepset-ai/haystack/commit/35788a2d060d58b901b8d10237f23273399998e5)  | Test    | 3.236s    | 0.185s   | **3.420s**     |
| This PR       | Minimal | 2.098s    | 0.036s   | 2.134s     |
| This PR       | Test    | 2.265s    | 0.050s   | **2.316s**     |


### Next Steps
I believe that these results can encourage discussion.

- Should we rename `LazyImport` or transform it into a truly lazy importer (difficult if we want to keep its API unchanged)?
- Do we want to use *real* lazy imports in the `__init__.py` of packages?