@anakin87 2025-02-20
closing this draft and continuing the work in https://github.com/deepset-ai/haystack/pull/8878